fix(db): idempotent ALTER TABLE ADD COLUMN migrations (#112)
Extract migration runner into its own submodule and wrap every
ALTER TABLE ADD COLUMN / RENAME COLUMN in a pragma_table_info check
so partial prior state does not break resumption. Fixes the #112
failure mode where schema_version was rewound below the actual
schema state, causing V5 to fail with 'duplicate column name'.

Also moves the two new migration tests into migration_tests.rs to
keep tests.rs under the 1000-line file-size gate.
