+++ crates/rs-core/src/db/mod.rs
@@ -12,6 +12,9 @@ pub use templates::*;
 mod v2;
 pub use v2::*;
 
+mod migrations;
+pub use migrations::{MAX_SCHEMA_VERSION, run_migrations};
+
 pub mod upload;
 pub use upload::{
     list_recent_uploads, mark_upload_permanently_failed, pick_next_uploadable_chunk,
@@ -32,6 +35,9 @@ mod delivery_log_tests;
 
 #[cfg(test)]
 mod delivery_status_tests;
+
+#[cfg(test)]
+mod migration_tests;
 
 /// Create a SQLite connection pool.
 pub async fn create_pool(db_path: &Path) -> Result<SqlitePool> {
@@ -58,354 +64,6 @@ pub async fn create_memory_pool() -> Result<SqlitePool> {
         .await?;
     Ok(pool)
 }
-
-/// Maximum schema version. Must equal the highest version in the migration list.
-/// Tests assert that `run_migrations` reaches this exact value.
-pub const MAX_SCHEMA_VERSION: i32 = 17;
-
-/// Run database migrations.
-///
-/// Uses a version tracking table to support incremental schema changes.
-/// Wrapped in a transaction so partial failures don't leave the DB inconsistent.
-pub async fn run_migrations(pool: &SqlitePool) -> Result<()> {
-    sqlx::query("CREATE TABLE IF NOT EXISTS schema_version (version INTEGER PRIMARY KEY)")
-        .execute(pool)
-        .await?;
-
-    let current: i32 = sqlx::query("SELECT COALESCE(MAX(version), 0) as v FROM schema_version")
-        .fetch_one(pool)
-        .await
-        .map(|r| r.get("v"))?;
-
-    let migrations: &[(i32, &str)] = &[
-        (1, MIGRATION_V1_SQL),
-        (2, MIGRATION_V2_SQL),
-        (3, MIGRATION_V3_SQL),
-        (4, MIGRATION_V4_SQL),
-        (5, MIGRATION_V5_SQL),
-        (6, MIGRATION_V6_SQL),
-        (7, MIGRATION_V7_SQL),
-        (8, MIGRATION_V8_SQL),
-        (9, MIGRATION_V9_SQL),
-        (10, MIGRATION_V10_SQL),
-        (11, MIGRATION_V11_SQL),
-        (12, MIGRATION_V12_SQL),
-        (13, MIGRATION_V13_SQL),
-        (14, MIGRATION_V14_SQL),
-        (15, MIGRATION_V15_SQL),
-        (16, MIGRATION_V16_SQL),
-        (17, MIGRATION_V17_SQL),
-    ];
-
-    for &(version, sql) in migrations {
-        if current < version {
-            let mut tx = pool.begin().await?;
-            for statement in sql.split(';') {
-                let trimmed = statement.trim();
-                if !trimmed.is_empty() {
-                    sqlx::query(trimmed).execute(&mut *tx).await?;
-                }
-            }
-            sqlx::query("INSERT OR REPLACE INTO schema_version (version) VALUES (?1)")
-                .bind(version)
-                .execute(&mut *tx)
-                .await?;
-            tx.commit().await?;
-        }
-    }
-
-    // Startup cleanup: delete old sent chunk records to keep the DB fast.
-    // Without this, CI runs accumulate 100K+ rows making startup take >30s.
-    let deleted: i64 = sqlx::query(
-        "DELETE FROM chunk_records WHERE sent = 1 AND created_at < datetime('now', '-1 hour')",
-    )
-    .execute(pool)
-    .await
-    .map(|r| r.rows_affected() as i64)
-    .unwrap_or(0);
-    if deleted > 0 {
-        tracing::info!("Cleaned {deleted} old chunk records from database");
-    }
-
-    Ok(())
-}
-
-const MIGRATION_V1_SQL: &str = r#"
-CREATE TABLE IF NOT EXISTS client_profile (
-    id        INTEGER PRIMARY KEY,
-    user_uuid TEXT NOT NULL UNIQUE
-);
-
-CREATE TABLE IF NOT EXISTS streaming_events (
-    id                   INTEGER PRIMARY KEY AUTOINCREMENT,
-    identifier           TEXT UNIQUE,
-    short_description    TEXT,
-    date_of_event        TEXT NOT NULL DEFAULT (datetime('now')),
-    server_ip            TEXT DEFAULT '',
-    received_bytes       INTEGER NOT NULL DEFAULT 0,
-    receiving_activated  INTEGER NOT NULL DEFAULT 0,
-    delivering_activated INTEGER NOT NULL DEFAULT 0
-);
-
-CREATE TABLE IF NOT EXISTS chunk_records (
-    id                 INTEGER PRIMARY KEY AUTOINCREMENT,
-    streaming_event_id INTEGER NOT NULL REFERENCES streaming_events(id) ON DELETE CASCADE,
-    chunk_file_path    TEXT NOT NULL,
-    data_size          INTEGER NOT NULL,
-    created_at         TEXT NOT NULL DEFAULT (datetime('now')),
-    md5                TEXT NOT NULL DEFAULT '',
-    in_process         INTEGER NOT NULL DEFAULT 0,
-    sent               INTEGER NOT NULL DEFAULT 0
-);
-
-CREATE INDEX IF NOT EXISTS idx_chunks_unsent ON chunk_records(streaming_event_id, sent, in_process)
-    WHERE sent = 0 AND in_process = 0
-"#;
-
-const MIGRATION_V2_SQL: &str = r#"
-CREATE TABLE IF NOT EXISTS endpoint_configs (
-    id             INTEGER PRIMARY KEY AUTOINCREMENT,
-    alias          TEXT NOT NULL UNIQUE,
-    service_type   TEXT NOT NULL CHECK(service_type IN ('YT_HLS','FB','YT_RTMP','VIMEO','INSTAGRAM','TEST_FILE')),
-    stream_key     TEXT NOT NULL DEFAULT '',
-    enabled        INTEGER NOT NULL DEFAULT 1,
-    position_last  INTEGER NOT NULL DEFAULT 0,
-    delivered_bytes INTEGER NOT NULL DEFAULT 0,
-    is_fast        INTEGER NOT NULL DEFAULT 0,
-    created_at     TEXT NOT NULL DEFAULT (datetime('now')),
-    updated_at     TEXT NOT NULL DEFAULT (datetime('now'))
-);
-
-CREATE TABLE IF NOT EXISTS event_endpoints (
-    event_id    INTEGER NOT NULL REFERENCES streaming_events(id) ON DELETE CASCADE,
-    endpoint_id INTEGER NOT NULL REFERENCES endpoint_configs(id) ON DELETE CASCADE,
-    PRIMARY KEY (event_id, endpoint_id)
-);
-
-CREATE TABLE IF NOT EXISTS delivery_instances (
-    id             INTEGER PRIMARY KEY AUTOINCREMENT,
-    hetzner_id     INTEGER NOT NULL UNIQUE,
-    name           TEXT NOT NULL,
-    ipv4           TEXT NOT NULL DEFAULT '',
-    status         TEXT NOT NULL DEFAULT 'creating' CHECK(status IN ('creating','running','stopping','deleted')),
-    server_type    TEXT NOT NULL DEFAULT 'cx23',
-    event_id       INTEGER REFERENCES streaming_events(id) ON DELETE SET NULL,
-    created_at     TEXT NOT NULL DEFAULT (datetime('now')),
-    last_health_at TEXT,
-    auth_token     TEXT NOT NULL DEFAULT ''
-);
-
-CREATE TABLE IF NOT EXISTS delivery_endpoint_status (
-    id              INTEGER PRIMARY KEY AUTOINCREMENT,
-    instance_id     INTEGER NOT NULL REFERENCES delivery_instances(id) ON DELETE CASCADE,
-    alias           TEXT NOT NULL,
-    alive           INTEGER NOT NULL DEFAULT 0,
-    buff_size_bytes INTEGER NOT NULL DEFAULT 0,
-    current_chunk_id INTEGER NOT NULL DEFAULT 0,
-    last_check_at   TEXT NOT NULL DEFAULT (datetime('now'))
-);
-
-CREATE TABLE IF NOT EXISTS youtube_oauth (
-    id             INTEGER PRIMARY KEY DEFAULT 1,
-    access_token   TEXT NOT NULL DEFAULT '',
-    refresh_token  TEXT NOT NULL DEFAULT '',
-    token_uri      TEXT NOT NULL DEFAULT 'https://oauth2.googleapis.com/token',
-    client_id      TEXT NOT NULL DEFAULT '',
-    client_secret  TEXT NOT NULL DEFAULT '',
-    scopes         TEXT NOT NULL DEFAULT '',
-    expires_at     TEXT
-);
-
-"#;
-
-const MIGRATION_V3_SQL: &str = r#"
-DROP TABLE IF EXISTS scheduled_streams;
-
-CREATE TABLE IF NOT EXISTS streaming_events_new (
-    id                   INTEGER PRIMARY KEY AUTOINCREMENT,
-    name                 TEXT NOT NULL UNIQUE,
-    received_bytes       INTEGER NOT NULL DEFAULT 0,
-    receiving_activated  INTEGER NOT NULL DEFAULT 0,
-    delivering_activated INTEGER NOT NULL DEFAULT 0
-);
-
-INSERT INTO streaming_events_new (id, name, received_bytes, receiving_activated, delivering_activated)
-    SELECT id, COALESCE(identifier, 'Event-' || id), received_bytes, receiving_activated, delivering_activated
-    FROM streaming_events;
-
-DROP TABLE streaming_events;
-
-ALTER TABLE streaming_events_new RENAME TO streaming_events
-"#;
-
-const MIGRATION_V4_SQL: &str = r#"
-CREATE UNIQUE INDEX IF NOT EXISTS idx_delivery_endpoint_status_instance_alias
-    ON delivery_endpoint_status(instance_id, alias);
-
-PRAGMA foreign_keys = OFF;
-
-CREATE TABLE IF NOT EXISTS delivery_instances_new (
-    id             INTEGER PRIMARY KEY AUTOINCREMENT,
-    hetzner_id     INTEGER NOT NULL UNIQUE,
-    name           TEXT NOT NULL,
-    ipv4           TEXT NOT NULL DEFAULT '',
-    status         TEXT NOT NULL DEFAULT 'creating' CHECK(status IN ('creating','running','stopping','deleted','failed')),
-    server_type    TEXT NOT NULL DEFAULT 'cx23',
-    event_id       INTEGER REFERENCES streaming_events(id) ON DELETE SET NULL,
-    created_at     TEXT NOT NULL DEFAULT (datetime('now')),
-    last_health_at TEXT
-);
-
-INSERT INTO delivery_instances_new (id, hetzner_id, name, ipv4, status, server_type, event_id, created_at, last_health_at)
-    SELECT id, hetzner_id, name, ipv4, status, server_type, event_id, created_at, last_health_at
-    FROM delivery_instances;
-
-DROP TABLE delivery_instances;
-
-ALTER TABLE delivery_instances_new RENAME TO delivery_instances;
-
-PRAGMA foreign_keys = ON
-"#;
-
-const MIGRATION_V5_SQL: &str = r#"
-ALTER TABLE delivery_instances ADD COLUMN auth_token TEXT NOT NULL DEFAULT ''
-"#;
-
-const MIGRATION_V6_SQL: &str = r#"
-ALTER TABLE chunk_records ADD COLUMN sent_at TEXT;
-ALTER TABLE delivery_endpoint_status ADD COLUMN bytes_processed_total INTEGER NOT NULL DEFAULT 0
-"#;
-
-const MIGRATION_V7_SQL: &str = r#"
-ALTER TABLE delivery_endpoint_status RENAME COLUMN buff_size_bytes TO chunks_processed
-"#;
-
-const MIGRATION_V8_SQL: &str = r#"
-ALTER TABLE chunk_records ADD COLUMN sequence_number INTEGER NOT NULL DEFAULT 0;
-
-UPDATE chunk_records SET sequence_number = (
-    SELECT COUNT(*) FROM chunk_records c2
-    WHERE c2.streaming_event_id = chunk_records.streaming_event_id
-    AND c2.id <= chunk_records.id
-);
-
-CREATE INDEX idx_chunks_event_sequence ON chunk_records(streaming_event_id, sequence_number)
-"#;
-
-const MIGRATION_V9_SQL: &str = r#"
-ALTER TABLE streaming_events ADD COLUMN cache_delay_secs INTEGER
-"#;
-
-const MIGRATION_V10_SQL: &str = r#"
-ALTER TABLE chunk_records ADD COLUMN chunk_format TEXT NOT NULL DEFAULT 'ts'
-"#;
-
-const MIGRATION_V11_SQL: &str = r#"
-ALTER TABLE chunk_records ADD COLUMN duration_ms INTEGER NOT NULL DEFAULT 0
-"#;
-
-const MIGRATION_V12_SQL: &str = r#"
-CREATE TABLE IF NOT EXISTS event_templates (
-    id               INTEGER PRIMARY KEY AUTOINCREMENT,
-    name             TEXT NOT NULL UNIQUE,
-    cache_delay_secs INTEGER
-);
-
-CREATE TABLE IF NOT EXISTS template_endpoints (
-    template_id INTEGER NOT NULL REFERENCES event_templates(id) ON DELETE CASCADE,
-    endpoint_id INTEGER NOT NULL REFERENCES endpoint_configs(id) ON DELETE CASCADE,
-    PRIMARY KEY (template_id, endpoint_id)
-);
-
-ALTER TABLE streaming_events ADD COLUMN created_from TEXT
-"#;
-
-const MIGRATION_V13_SQL: &str = r#"
-CREATE TABLE IF NOT EXISTS delivery_logs (
-    id          INTEGER PRIMARY KEY AUTOINCREMENT,
-    instance_id INTEGER NOT NULL,
-    event_id    INTEGER,
-    captured_at TEXT NOT NULL DEFAULT (datetime('now')),
-    log_text    TEXT NOT NULL
-);
-
-CREATE TABLE IF NOT EXISTS delivery_restart_log (
-    id            INTEGER PRIMARY KEY AUTOINCREMENT,
-    instance_id   INTEGER NOT NULL,
-    event_id      INTEGER,
-    alias         TEXT NOT NULL,
-    timestamp_ms  INTEGER NOT NULL,
-    chunk_id      INTEGER NOT NULL,
-    lifetime_secs INTEGER NOT NULL,
-    reason        TEXT NOT NULL,
-    stderr_tail   TEXT,
-    backoff_secs  INTEGER NOT NULL
-);
-
-CREATE INDEX idx_delivery_restart_log_instance
-    ON delivery_restart_log(instance_id);
-
-CREATE INDEX idx_delivery_logs_instance
-    ON delivery_logs(instance_id)
-"#;
-
-const MIGRATION_V14_SQL: &str = r#"
-ALTER TABLE streaming_events ADD COLUMN rescue_video_url TEXT
-"#;
-
-const MIGRATION_V15_SQL: &str = r#"
-ALTER TABLE event_templates ADD COLUMN rescue_video_url TEXT
-"#;
-
-// V16: widen the delivery_instances.status CHECK constraint to allow the
-// new orchestrator phases (booting, initializing, delivering). Without
-// this, writes from poll_and_init to "booting" fail with
-//   CHECK constraint failed: status IN ('creating','running',...)
-// which then sets the instance to "failed" and leaves the operator
-// staring at a broken dashboard.
-//
-// SQLite doesn't support ALTER TABLE ... DROP CONSTRAINT so we have to
-// recreate the table, copying rows over.
-const MIGRATION_V16_SQL: &str = r#"
-CREATE TABLE delivery_instances_v16 (
-    id              INTEGER PRIMARY KEY AUTOINCREMENT,
-    hetzner_id      INTEGER NOT NULL UNIQUE,
-    name            TEXT NOT NULL,
-    ipv4            TEXT NOT NULL,
-    status          TEXT NOT NULL DEFAULT 'creating' CHECK(status IN (
-        'creating', 'running', 'stopping', 'deleted', 'failed',
-        'booting', 'initializing', 'delivering'
-    )),
-    server_type     TEXT NOT NULL,
-    event_id        INTEGER REFERENCES streaming_events(id) ON DELETE SET NULL,
-    created_at      TEXT NOT NULL DEFAULT (datetime('now')),
-    last_health_at  TEXT,
-    auth_token      TEXT NOT NULL DEFAULT ''
-);
-
-INSERT INTO delivery_instances_v16
-    (id, hetzner_id, name, ipv4, status, server_type, event_id, created_at, last_health_at, auth_token)
-SELECT
-    id, hetzner_id, name, ipv4, status, server_type, event_id, created_at, last_health_at, auth_token
-FROM delivery_instances;
-
-DROP TABLE delivery_instances;
-ALTER TABLE delivery_instances_v16 RENAME TO delivery_instances
-"#;
-
-const MIGRATION_V17_SQL: &str = r#"
-ALTER TABLE chunk_records ADD COLUMN upload_attempts INTEGER NOT NULL DEFAULT 0;
-ALTER TABLE chunk_records ADD COLUMN upload_first_attempt_at INTEGER;
-ALTER TABLE chunk_records ADD COLUMN upload_completed_at INTEGER;
-ALTER TABLE chunk_records ADD COLUMN upload_duration_ms INTEGER;
-ALTER TABLE chunk_records ADD COLUMN upload_last_error TEXT;
-ALTER TABLE chunk_records ADD COLUMN upload_next_retry_at INTEGER;
-ALTER TABLE chunk_records ADD COLUMN upload_failed_permanently INTEGER NOT NULL DEFAULT 0;
-
-CREATE INDEX IF NOT EXISTS idx_chunks_upload_queue
-  ON chunk_records(upload_failed_permanently, sent, in_process, upload_next_retry_at, id)
-  WHERE sent = 0 AND in_process = 0 AND upload_failed_permanently = 0
-"#;
 
 // --- Client Profile ---
